Move ModuleReferenceEvent and related types to separate file
