+++ backend-rust/src/graphql_api.rs
@@ -7,6 +7,7 @@ mod account_metrics;
 mod baker;
 mod block;
 mod block_metrics;
+mod module_reference_event;
 mod transaction;
 mod transaction_metrics;
 
@@ -20,10 +21,10 @@ macro_rules! todo_api {
 pub(crate) use todo_api;
 
 use crate::{
-    address::{AccountAddress, ContractAddress, ContractIndex},
+    address::{AccountAddress, ContractIndex},
     scalar_types::{
-        AccountIndex, Amount, BigInteger, BlockHeight, DateTime, ModuleReference, TimeSpan,
-        TransactionHash, TransactionIndex,
+        AccountIndex, Amount, BigInteger, BlockHeight, DateTime, TimeSpan, TransactionHash,
+        TransactionIndex,
     },
     transaction_event::{
         delegation::{BakerDelegationTarget, DelegationTarget, PassiveDelegationTarget},
@@ -48,8 +49,7 @@ use bigdecimal::BigDecimal;
 use block::Block;
 use chrono::Duration;
 use concordium_rust_sdk::{
-    base::contracts_common::schema::{VersionedModuleSchema, VersionedSchemaError},
-    id::types as sdk_types,
+    base::contracts_common::schema::VersionedSchemaError, id::types as sdk_types,
 };
 use futures::prelude::*;
 use prometheus_client::registry::Registry;
@@ -125,6 +125,7 @@ pub struct Query(
     account_metrics::QueryAccountMetrics,
     transaction_metrics::QueryTransactionMetrics,
     block_metrics::QueryBlockMetrics,
+    module_reference_event::QueryModuleReferenceEvent,
 );
 
 pub struct Service {
@@ -788,48 +789,6 @@ impl BaseQuery {
         }
 
         Ok(connection)
-    }
-
-    async fn module_reference_event<'a>(
-        &self,
-        ctx: &Context<'a>,
-        module_reference: String,
-    ) -> ApiResult<ModuleReferenceEvent> {
-        let pool = get_pool(ctx)?;
-
-        let row = sqlx::query!(
-            r#"SELECT
-                blocks.height as block_height,
-                smart_contract_modules.transaction_index,
-                schema as display_schema,
-                blocks.slot_time as block_slot_time,
-                transactions.hash as transaction_hash,
-                accounts.address as sender
-            FROM smart_contract_modules
-            JOIN transactions ON smart_contract_modules.transaction_index = transactions.index
-            JOIN blocks ON transactions.block_height = blocks.height
-            JOIN accounts ON transactions.sender = accounts.index
-            WHERE module_reference = $1"#,
-            module_reference
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)?;
-
-        let display_schema = row
-            .display_schema
-            .as_ref()
-            .map(|s| VersionedModuleSchema::new(s, &None).map(|schema| schema.to_string()))
-            .transpose()?;
-
-        Ok(ModuleReferenceEvent {
-            module_reference,
-            sender: row.sender.into(),
-            block_height: row.block_height,
-            transaction_hash: row.transaction_hash,
-            block_slot_time: row.block_slot_time,
-            display_schema,
-        })
     }
 }
 
@@ -2142,196 +2101,6 @@ impl SearchResult {
     ) -> ApiResult<connection::Connection<String, NodeStatus>> {
         todo_api!()
     }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct ModuleReferenceEvent {
-    module_reference: ModuleReference,
-    sender:           AccountAddress,
-    block_height:     BlockHeight,
-    transaction_hash: TransactionHash,
-    block_slot_time:  DateTime,
-    display_schema:   Option<String>,
-    // TODO: linkedContracts(skip: Int take: Int): LinkedContractsCollectionSegment
-}
-#[ComplexObject]
-impl ModuleReferenceEvent {
-    async fn module_reference_reject_events(
-        &self,
-        ctx: &Context<'_>,
-        skip: Option<u64>,
-        take: Option<u64>,
-    ) -> ApiResult<ModuleReferenceRejectEventsCollectionSegment> {
-        let pool = get_pool(ctx)?;
-        let config = get_config(ctx)?;
-        let min_index = i64::try_from(skip.unwrap_or(0))?;
-        let limit = i64::try_from(
-            take.map_or(config.module_reference_reject_events_collection_limit, |t| {
-                config.module_reference_reject_events_collection_limit.min(t)
-            }),
-        )?;
-
-        let mut items = sqlx::query_as!(
-            ModuleReferenceRejectEvent,
-            r#"SELECT
-                module_reference,
-                transactions.reject as "reject: sqlx::types::Json<TransactionRejectReason>",
-                transactions.block_height,
-                transactions.hash as transaction_hash,
-                blocks.slot_time as block_slot_time
-            FROM rejected_smart_contract_module_transactions
-                JOIN transactions ON transaction_index = transactions.index
-                JOIN blocks ON blocks.height = transactions.block_height
-            WHERE module_reference = $1
-                AND rejected_smart_contract_module_transactions.index >= $2
-            LIMIT $3
-        "#,
-            self.module_reference,
-            min_index,
-            limit + 1
-        )
-        .fetch_all(pool)
-        .await?;
-
-        // Determine if there is a next page by checking if we got more than `limit`
-        // rows.
-        let has_next_page = items.len() > limit as usize;
-        // If there is a next page, remove the extra row used for pagination detection.
-        if has_next_page {
-            items.pop();
-        }
-        let has_previous_page = min_index > 0;
-        Ok(ModuleReferenceRejectEventsCollectionSegment {
-            page_info: CollectionSegmentInfo {
-                has_next_page,
-                has_previous_page,
-            },
-            total_count: items.len().try_into()?,
-            items,
-        })
-    }
-
-    async fn module_reference_contract_link_events(
-        &self,
-        ctx: &Context<'_>,
-        skip: Option<u64>,
-        take: Option<u64>,
-    ) -> ApiResult<ModuleReferenceContractLinkEventsCollectionSegment> {
-        let pool = get_pool(ctx)?;
-        let config = get_config(ctx)?;
-        let min_index = i64::try_from(skip.unwrap_or(0))?;
-        let limit = i64::try_from(
-            take.map_or(config.module_reference_contract_link_events_collection_limit, |t| {
-                config.module_reference_contract_link_events_collection_limit.min(t)
-            }),
-        )?;
-
-        let mut items = sqlx::query_as!(
-            ModuleReferenceContractLinkEvent,
-            r#"SELECT
-                link_action as "link_action: ModuleReferenceContractLinkAction",
-                contract_index,
-                contract_sub_index,
-                transactions.hash as transaction_hash,
-                blocks.slot_time as block_slot_time
-            FROM link_smart_contract_module_transactions
-                JOIN transactions ON transaction_index = transactions.index
-                JOIN blocks ON blocks.height = transactions.block_height
-            WHERE module_reference = $1
-                AND link_smart_contract_module_transactions.index >= $2
-            LIMIT $3
-        "#,
-            self.module_reference,
-            min_index,
-            limit + 1
-        )
-        .fetch_all(pool)
-        .await?;
-
-        // Determine if there is a next page by checking if we got more than `limit`
-        // rows.
-        let has_next_page = items.len() > limit as usize;
-        // If there is a next page, remove the extra row used for pagination detection.
-        if has_next_page {
-            items.pop();
-        }
-        let has_previous_page = min_index > 0;
-
-        Ok(ModuleReferenceContractLinkEventsCollectionSegment {
-            page_info: CollectionSegmentInfo {
-                has_next_page,
-                has_previous_page,
-            },
-            total_count: items.len().try_into()?,
-            items,
-        })
-    }
-}
-
-#[derive(SimpleObject)]
-struct ModuleReferenceRejectEventsCollectionSegment {
-    page_info:   CollectionSegmentInfo,
-    items:       Vec<ModuleReferenceRejectEvent>,
-    total_count: u64,
-}
-
-#[derive(SimpleObject)]
-#[graphql(complex)]
-struct ModuleReferenceRejectEvent {
-    module_reference: ModuleReference,
-    #[graphql(skip)]
-    reject:           Option<sqlx::types::Json<TransactionRejectReason>>,
-    block_height:     BlockHeight,
-    transaction_hash: TransactionHash,
-    block_slot_time:  DateTime,
-}
-#[ComplexObject]
-impl ModuleReferenceRejectEvent {
-    async fn rejected_event(&self) -> ApiResult<&TransactionRejectReason> {
-        if let Some(sqlx::types::Json(reason)) = self.reject.as_ref() {
-            Ok(reason)
-        } else {
-            Err(ApiError::InternalError(
-                "ModuleReferenceRejectEvent: No reject reason found".to_string(),
-            ))
-        }
-    }
-}
-
-#[derive(SimpleObject)]
-#[graphql(complex)]
-struct ModuleReferenceContractLinkEvent {
-    block_slot_time:    DateTime,
-    transaction_hash:   TransactionHash,
-    link_action:        ModuleReferenceContractLinkAction,
-    #[graphql(skip)]
-    contract_index:     i64,
-    #[graphql(skip)]
-    contract_sub_index: i64,
-}
-#[ComplexObject]
-impl ModuleReferenceContractLinkEvent {
-    async fn contract_address(&self) -> ApiResult<ContractAddress> {
-        ContractAddress::new(self.contract_index, self.contract_sub_index)
-    }
-}
-
-#[derive(Enum, Clone, Copy, PartialEq, Eq, sqlx::Type)]
-#[sqlx(type_name = "module_reference_contract_link_action")]
-pub enum ModuleReferenceContractLinkAction {
-    Added,
-    Removed,
-}
-
-/// A segment of a collection.
-#[derive(SimpleObject)]
-struct ModuleReferenceContractLinkEventsCollectionSegment {
-    /// Information to aid in pagination.
-    page_info:   CollectionSegmentInfo,
-    /// A flattened list of the items.
-    items:       Vec<ModuleReferenceContractLinkEvent>,
-    total_count: u64,
 }
 
 #[derive(Enum, Clone, Copy, PartialEq, Eq)]
